requirements.txt
Django==3.2
Pillow==8.2.0
psycopg2==2.8.6
python-telegram-bot==13.4.1
